+++ backend/app/routers/idioms.py
@@ -1,0 +1,108 @@
+from typing import Annotated
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, HTTPException, Query
+from sqlalchemy.orm import Session
+
+from app import database
+from app.models.idioms import IdiomModel
+from app.schemas.idioms import IdiomCreate, IdiomSchema, IdiomUpdate
+
+SessionDep = Annotated[Session, Depends(database.get_session)]
+
+router = APIRouter(prefix="/idioms", tags=["idioms"])
+
+
+@router.get("/", response_model=list[IdiomSchema])
+async def get_idioms(
+    db: SessionDep,
+    page: int = 1,
+    limit: Annotated[int, Query(le=50)] = 50,
+    search: Annotated[str, Query()] = "",
+) -> list[IdiomSchema]:
+    search = search.strip()
+    offset = (page - 1) * limit
+
+    if not search:
+        return [
+            IdiomSchema.model_validate(idiom)
+            for idiom in db.query(IdiomModel)
+            .order_by(IdiomModel.text.asc())
+            .limit(limit)
+            .offset(offset)
+            .all()
+        ]
+    else:
+        return [
+            IdiomSchema.model_validate(idiom)
+            for idiom in db.query(IdiomModel)
+            .filter(IdiomModel.text.ilike(f"%{search}%"))
+            .order_by(IdiomModel.text.asc())
+            .limit(limit)
+            .offset(offset)
+            .all()
+        ]
+
+
+@router.get("/favorites", response_model=list[IdiomSchema])
+async def get_favorite_idioms(
+    db: SessionDep,
+    page: int = 1,
+    limit: Annotated[int, Query(le=50)] = 50,
+) -> list[IdiomSchema]:
+    offset = (page - 1) * limit
+
+    return [
+        IdiomSchema.model_validate(idiom)
+        for idiom in db.query(IdiomModel)
+        .filter(IdiomModel.favorite)
+        .order_by(IdiomModel.text.asc())
+        .limit(limit)
+        .offset(offset)
+        .all()
+    ]
+
+
+@router.post("/", status_code=201)
+async def post_idioms(db: SessionDep, payload: list[IdiomCreate]) -> None:
+    idioms = [IdiomModel(**idiom.model_dump()) for idiom in payload]
+    db.add_all(idioms)
+    db.commit()
+
+
+@router.post("/{id}/upvote", response_model=IdiomSchema)
+async def upvote_idiom(db: SessionDep, id: UUID) -> IdiomSchema:
+    idiom = db.query(IdiomModel).filter(IdiomModel.id == id).first()
+    if not idiom:
+        raise HTTPException(status_code=404, detail="Idiom not found")
+    idiom.upvotes += 1
+    db.add(idiom)
+    db.commit()
+    db.refresh(idiom)
+    return idiom
+
+
+@router.post("/{id}/downvote", response_model=IdiomSchema)
+async def downvote_idiom(db: SessionDep, id: UUID) -> IdiomSchema:
+    idiom = db.query(IdiomModel).filter(IdiomModel.id == id).first()
+    if not idiom:
+        raise HTTPException(status_code=404, detail="Idiom not found")
+    idiom.downvotes += 1
+    db.add(idiom)
+    db.commit()
+    db.refresh(idiom)
+    return idiom
+
+
+@router.patch("/{id}", response_model=IdiomSchema)
+async def update_idiom(db: SessionDep, id: UUID, payload: IdiomUpdate) -> IdiomSchema:
+    idiom = db.query(IdiomModel).filter(IdiomModel.id == id).first()
+    if not idiom:
+        raise HTTPException(status_code=404, detail="Idiom not found")
+    idiom_update = payload.model_dump(exclude_unset=True)
+    for key, value in idiom_update.items():
+        setattr(idiom, key, value)
+    db.add(idiom)
+    db.commit()
+    db.refresh(idiom)
+    return idiom
